+++ src/lib/navbar/collapse/collapse.props.ts
@@ -1,6 +1,5 @@
 import React from 'react'
 
-interface CollapseProps extends React.HTMLAttributes<HTMLElement> {
-}
+interface CollapseProps extends React.HTMLAttributes<HTMLElement> {}
 
 export default CollapseProps