+++ src/lib/navbar/item/item.props.ts
@@ -1,7 +1,8 @@
 import React from 'react'
 import { AsProp } from '../../helper'
 
-interface ItemProps extends React.AnchorHTMLAttributes<HTMLAnchorElement> , AsProp {
-}
+interface ItemProps
+  extends React.AnchorHTMLAttributes<HTMLAnchorElement>,
+  AsProp {}
 
 export default ItemProps