+++ src/lib/navbar/items/items.props.ts
@@ -2,7 +2,8 @@ import React from 'react'
 import { AsProp } from '../../helper'
 
 interface ItemsProps
-  extends AsProp, Omit<React.AnchorHTMLAttributes<HTMLAnchorElement>, 'title'> {
+  extends AsProp,
+  Omit<React.AnchorHTMLAttributes<HTMLAnchorElement>, 'title'> {
   title?: React.ReactNode
 }