+++ src/lib/navbar/logo/logo.props.ts
@@ -1,7 +1,8 @@
 import React from 'react'
 import { AsProp } from '../../helper'
 
-interface LogoProps extends React.AnchorHTMLAttributes<HTMLAnchorElement> , AsProp {
-}
+interface LogoProps
+  extends React.AnchorHTMLAttributes<HTMLAnchorElement>,
+  AsProp {}
 
 export default LogoProps